atualizando jsons e add valores inteiros nas tabelas
