--- src/controllers/abstencao_atualizado_json.js
@@ -1,941 +1,0 @@
-const data = {
-  comparecimento_abstencao: [
-    {
-      municipio: "SÃO PAULO",
-      QT_ABSTENCAO: 2632592,
-      QT_COMPARECIMENTO: 6354095,
-      faixa_etaria: [
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "100 anos ou mais",
-          qt_aptos: 3170,
-          qt_abstencao: 3153,
-          qt_comparecimento: 17,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "16 anos",
-          qt_aptos: 64550,
-          qt_abstencao: 12307,
-          qt_comparecimento: 52243,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "17 anos",
-          qt_aptos: 120976,
-          qt_abstencao: 29693,
-          qt_comparecimento: 91283,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "18 anos",
-          qt_aptos: 155430,
-          qt_abstencao: 41706,
-          qt_comparecimento: 113724,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "19 anos",
-          qt_aptos: 168729,
-          qt_abstencao: 48046,
-          qt_comparecimento: 120683,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "20 anos",
-          qt_aptos: 171363,
-          qt_abstencao: 49625,
-          qt_comparecimento: 121738,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "21 a 24 anos",
-          qt_aptos: 685485,
-          qt_abstencao: 196665,
-          qt_comparecimento: 488820,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "25 a 29 anos",
-          qt_aptos: 841102,
-          qt_abstencao: 239262,
-          qt_comparecimento: 601840,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "30 a 34 anos",
-          qt_aptos: 934087,
-          qt_abstencao: 253662,
-          qt_comparecimento: 680425,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "35 a 39 anos",
-          qt_aptos: 986448,
-          qt_abstencao: 248933,
-          qt_comparecimento: 737515,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "40 a 44 anos",
-          qt_aptos: 879529,
-          qt_abstencao: 198748,
-          qt_comparecimento: 680781,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "45 a 49 anos",
-          qt_aptos: 786324,
-          qt_abstencao: 156192,
-          qt_comparecimento: 630132,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "50 a 54 anos",
-          qt_aptos: 744397,
-          qt_abstencao: 133343,
-          qt_comparecimento: 611054,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "55 a 59 anos",
-          qt_aptos: 654639,
-          qt_abstencao: 115235,
-          qt_comparecimento: 539404,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "60 a 64 anos",
-          qt_aptos: 552420,
-          qt_abstencao: 108090,
-          qt_comparecimento: 444330,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "65 a 69 anos",
-          qt_aptos: 410792,
-          qt_abstencao: 136004,
-          qt_comparecimento: 274788,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "70 a 74 anos",
-          qt_aptos: 292576,
-          qt_abstencao: 195676,
-          qt_comparecimento: 96900,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "75 a 79 anos",
-          qt_aptos: 198915,
-          qt_abstencao: 161366,
-          qt_comparecimento: 37549,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "80 a 84 anos",
-          qt_aptos: 132759,
-          qt_abstencao: 119430,
-          qt_comparecimento: 13329,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "85 a 89 anos",
-          qt_aptos: 83992,
-          qt_abstencao: 80431,
-          qt_comparecimento: 3561,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "90 a 94 anos",
-          qt_aptos: 67044,
-          qt_abstencao: 66383,
-          qt_comparecimento: 661,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "95 a 99 anos",
-          qt_aptos: 30196,
-          qt_abstencao: 30113,
-          qt_comparecimento: 83,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "Inválido",
-          qt_aptos: 21764,
-          qt_abstencao: 8529,
-          qt_comparecimento: 13235,
-        },
-      ],
-      estado_civil: [
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "CASADO",
-          qt_aptos: 2993526,
-          qt_abstencao: 924778,
-          qt_comparecimento: 2068748,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "DIVORCIADO",
-          qt_aptos: 371269,
-          qt_abstencao: 95341,
-          qt_comparecimento: 275928,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "NÃO INFORMADO",
-          qt_aptos: 8855,
-          qt_abstencao: 5382,
-          qt_comparecimento: 3473,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "SEPARADO JUDICIALMENTE",
-          qt_aptos: 104954,
-          qt_abstencao: 39266,
-          qt_comparecimento: 65688,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "SOLTEIRO",
-          qt_aptos: 5279639,
-          qt_abstencao: 1456160,
-          qt_comparecimento: 3823479,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "VIÚVO",
-          qt_aptos: 228444,
-          qt_abstencao: 111665,
-          qt_comparecimento: 116779,
-        },
-      ],
-      grau_escolaridade: [
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ANALFABETO",
-          qt_aptos: 178351,
-          qt_abstencao: 112449,
-          qt_comparecimento: 65902,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL COMPLETO",
-          qt_aptos: 669859,
-          qt_abstencao: 247027,
-          qt_comparecimento: 422832,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL INCOMPLETO",
-          qt_aptos: 1597042,
-          qt_abstencao: 556498,
-          qt_comparecimento: 1040544,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO MÉDIO COMPLETO",
-          qt_aptos: 2492188,
-          qt_abstencao: 603320,
-          qt_comparecimento: 1888868,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO MÉDIO INCOMPLETO",
-          qt_aptos: 1615763,
-          qt_abstencao: 462652,
-          qt_comparecimento: 1153111,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "LÊ E ESCREVE",
-          qt_aptos: 382237,
-          qt_abstencao: 188194,
-          qt_comparecimento: 194043,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "SUPERIOR COMPLETO",
-          qt_aptos: 1444126,
-          qt_abstencao: 314426,
-          qt_comparecimento: 1129700,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "SUPERIOR INCOMPLETO",
-          qt_aptos: 607121,
-          qt_abstencao: 148026,
-          qt_comparecimento: 459095,
-        },
-      ],
-    },
-    {
-      municipio: "SÃO JOSÉ DOS CAMPOS",
-      QT_ABSTENCAO: 145210,
-      QT_COMPARECIMENTO: 395291,
-      faixa_etaria: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "100 anos ou mais",
-          qt_aptos: 75,
-          qt_abstencao: 75,
-          qt_comparecimento: 0,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "16 anos",
-          qt_aptos: 5040,
-          qt_abstencao: 999,
-          qt_comparecimento: 4041,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "17 anos",
-          qt_aptos: 8256,
-          qt_abstencao: 2280,
-          qt_comparecimento: 5976,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "18 anos",
-          qt_aptos: 9979,
-          qt_abstencao: 2808,
-          qt_comparecimento: 7171,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "19 anos",
-          qt_aptos: 10234,
-          qt_abstencao: 3217,
-          qt_comparecimento: 7017,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "20 anos",
-          qt_aptos: 10410,
-          qt_abstencao: 3328,
-          qt_comparecimento: 7082,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "21 a 24 anos",
-          qt_aptos: 41742,
-          qt_abstencao: 13326,
-          qt_comparecimento: 28416,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "25 a 29 anos",
-          qt_aptos: 54621,
-          qt_abstencao: 16098,
-          qt_comparecimento: 38523,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "30 a 34 anos",
-          qt_aptos: 58468,
-          qt_abstencao: 15362,
-          qt_comparecimento: 43106,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "35 a 39 anos",
-          qt_aptos: 62186,
-          qt_abstencao: 14581,
-          qt_comparecimento: 47605,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "40 a 44 anos",
-          qt_aptos: 53045,
-          qt_abstencao: 11037,
-          qt_comparecimento: 42008,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "45 a 49 anos",
-          qt_aptos: 46897,
-          qt_abstencao: 8590,
-          qt_comparecimento: 38307,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "50 a 54 anos",
-          qt_aptos: 43936,
-          qt_abstencao: 7306,
-          qt_comparecimento: 36630,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "55 a 59 anos",
-          qt_aptos: 39384,
-          qt_abstencao: 6173,
-          qt_comparecimento: 33211,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "60 a 64 anos",
-          qt_aptos: 33316,
-          qt_abstencao: 5790,
-          qt_comparecimento: 27526,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "65 a 69 anos",
-          qt_aptos: 24571,
-          qt_abstencao: 6894,
-          qt_comparecimento: 17677,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "70 a 74 anos",
-          qt_aptos: 16092,
-          qt_abstencao: 9476,
-          qt_comparecimento: 6616,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "75 a 79 anos",
-          qt_aptos: 9296,
-          qt_abstencao: 6951,
-          qt_comparecimento: 2345,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "80 a 84 anos",
-          qt_aptos: 5445,
-          qt_abstencao: 4765,
-          qt_comparecimento: 680,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "85 a 89 anos",
-          qt_aptos: 3027,
-          qt_abstencao: 2854,
-          qt_comparecimento: 173,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "90 a 94 anos",
-          qt_aptos: 1889,
-          qt_abstencao: 1861,
-          qt_comparecimento: 28,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "95 a 99 anos",
-          qt_aptos: 823,
-          qt_abstencao: 822,
-          qt_comparecimento: 1,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "Inválido",
-          qt_aptos: 1769,
-          qt_abstencao: 617,
-          qt_comparecimento: 1152,
-        },
-      ],
-      estado_civil: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "CASADO",
-          qt_aptos: 251759,
-          qt_abstencao: 57428,
-          qt_comparecimento: 194331,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "DIVORCIADO",
-          qt_aptos: 33764,
-          qt_abstencao: 8172,
-          qt_comparecimento: 25592,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "NÃO INFORMADO",
-          qt_aptos: 17,
-          qt_abstencao: 15,
-          qt_comparecimento: 2,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "SEPARADO JUDICIALMENTE",
-          qt_aptos: 9137,
-          qt_abstencao: 2803,
-          qt_comparecimento: 6334,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "SOLTEIRO",
-          qt_aptos: 227978,
-          qt_abstencao: 69941,
-          qt_comparecimento: 158037,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "VIÚVO",
-          qt_aptos: 17846,
-          qt_abstencao: 6851,
-          qt_comparecimento: 10995,
-        },
-      ],
-      grau_escolaridade: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ANALFABETO",
-          qt_aptos: 7450,
-          qt_abstencao: 5293,
-          qt_comparecimento: 2157,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL COMPLETO",
-          qt_aptos: 36895,
-          qt_abstencao: 10764,
-          qt_comparecimento: 26131,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL INCOMPLETO",
-          qt_aptos: 80780,
-          qt_abstencao: 27278,
-          qt_comparecimento: 53502,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO MÉDIO COMPLETO",
-          qt_aptos: 201805,
-          qt_abstencao: 46308,
-          qt_comparecimento: 155497,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO MÉDIO INCOMPLETO",
-          qt_aptos: 53476,
-          qt_abstencao: 18629,
-          qt_comparecimento: 34847,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "LÊ E ESCREVE",
-          qt_aptos: 11406,
-          qt_abstencao: 6193,
-          qt_comparecimento: 5213,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "SUPERIOR COMPLETO",
-          qt_aptos: 114388,
-          qt_abstencao: 22258,
-          qt_comparecimento: 92130,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "SUPERIOR INCOMPLETO",
-          qt_aptos: 34301,
-          qt_abstencao: 8487,
-          qt_comparecimento: 25814,
-        },
-      ],
-    },
-  ],
-  maiores_abstencoes: [
-    {
-      municipio: "MOCOCA",
-      min_abstencao: 33.5187,
-    },
-    {
-      municipio: "BARRETOS",
-      min_abstencao: 33.4652,
-    },
-    {
-      municipio: "SANTOS",
-      min_abstencao: 33.0131,
-    },
-    {
-      municipio: "CAMPOS DO JORDÃO",
-      min_abstencao: 32.8543,
-    },
-    {
-      municipio: "FERNANDÓPOLIS",
-      min_abstencao: 32.6283,
-    },
-  ],
-  menores_abstencoes: [
-    {
-      municipio: "PEDRANÓPOLIS",
-      min_abstencao: 7.7826,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      min_abstencao: 8.1261,
-    },
-    {
-      municipio: "PIQUEROBI",
-      min_abstencao: 8.164,
-    },
-    {
-      municipio: "NOVA CASTILHO",
-      min_abstencao: 8.4553,
-    },
-    {
-      municipio: "JERIQUARA",
-      min_abstencao: 9.2411,
-    },
-  ],
-  min_abstencoes_jovens: [
-    {
-      municipio: "URU",
-      abstencao_jovens: 7.0513,
-    },
-    {
-      municipio: "PIQUEROBI",
-      abstencao_jovens: 7.5253,
-    },
-    {
-      municipio: "NOVA CASTILHO",
-      abstencao_jovens: 8.1081,
-    },
-    {
-      municipio: "TRABIJU",
-      abstencao_jovens: 8.1264,
-    },
-    {
-      municipio: "CRUZÁLIA",
-      abstencao_jovens: 9.0498,
-    },
-  ],
-  max_abstencoes_jovens: [
-    {
-      municipio: "IARAS",
-      abstencao_jovens: 40.9913,
-    },
-    {
-      municipio: "ILHA SOLTEIRA",
-      abstencao_jovens: 37.4632,
-    },
-    {
-      municipio: "REGISTRO",
-      abstencao_jovens: 36.767,
-    },
-    {
-      municipio: "BARRETOS",
-      abstencao_jovens: 36.0859,
-    },
-    {
-      municipio: "ROSANA",
-      abstencao_jovens: 35.3393,
-    },
-  ],
-  min_abstencoes_adultos: [
-    {
-      municipio: "PIQUEROBI",
-      abstencao_adultos: 4.359,
-    },
-    {
-      municipio: "PEDRANÓPOLIS",
-      abstencao_adultos: 5.259,
-    },
-    {
-      municipio: "MARAPOAMA",
-      abstencao_adultos: 6.1599,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_adultos: 6.1684,
-    },
-    {
-      municipio: "URU",
-      abstencao_adultos: 6.1916,
-    },
-  ],
-  max_abstencoes_adultos: [
-    {
-      municipio: "CAMPOS DO JORDÃO",
-      abstencao_adultos: 25.6667,
-    },
-    {
-      municipio: "BARRETOS",
-      abstencao_adultos: 25.2493,
-    },
-    {
-      municipio: "JUQUIÁ",
-      abstencao_adultos: 25.0963,
-    },
-    {
-      municipio: "ITANHAÉM",
-      abstencao_adultos: 24.7228,
-    },
-    {
-      municipio: "MOCOCA",
-      abstencao_adultos: 24.5299,
-    },
-  ],
-  min_abstencoes_idosos: [
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_idosos: 10.6109,
-    },
-    {
-      municipio: "PEDRANÓPOLIS",
-      abstencao_idosos: 11.6858,
-    },
-    {
-      municipio: "GASTÃO VIDIGAL",
-      abstencao_idosos: 13.0996,
-    },
-    {
-      municipio: "NOVA LUZITÂNIA",
-      abstencao_idosos: 13.3056,
-    },
-    {
-      municipio: "PRACINHA",
-      abstencao_idosos: 13.5135,
-    },
-  ],
-  max_abstencoes_idosos: [
-    {
-      municipio: "PIEDADE",
-      abstencao_idosos: 56.1114,
-    },
-    {
-      municipio: "BARIRI",
-      abstencao_idosos: 56.1066,
-    },
-    {
-      municipio: "ITÁPOLIS",
-      abstencao_idosos: 54.7757,
-    },
-    {
-      municipio: "IGARAPAVA",
-      abstencao_idosos: 54.7103,
-    },
-    {
-      municipio: "RIBEIRÃO GRANDE",
-      abstencao_idosos: 54.4872,
-    },
-  ],
-  min_abstencoes_analfabeto: [
-    {
-      municipio: "ASPÁSIA",
-      abstencao_analfabetos: 17.1053,
-    },
-    {
-      municipio: "JERIQUARA",
-      abstencao_analfabetos: 17.2727,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_analfabetos: 17.8571,
-    },
-    {
-      municipio: "GASTÃO VIDIGAL",
-      abstencao_analfabetos: 18.1818,
-    },
-    {
-      municipio: "GUARANI D'OESTE",
-      abstencao_analfabetos: 18.5185,
-    },
-  ],
-  max_abstencoes_analfabeto: [
-    {
-      municipio: "LINS",
-      abstencao_analfabetos: 85.0237,
-    },
-    {
-      municipio: "JALES",
-      abstencao_analfabetos: 84.6617,
-    },
-    {
-      municipio: "PIEDADE",
-      abstencao_analfabetos: 82.8142,
-    },
-    {
-      municipio: "TUPÃ",
-      abstencao_analfabetos: 82.4598,
-    },
-    {
-      municipio: "PIRAPOZINHO",
-      abstencao_analfabetos: 81.0345,
-    },
-  ],
-  min_abstencoes_medio_completo: [
-    {
-      municipio: "ASPÁSIA",
-      abstencao_medio_completo: 17.1053,
-    },
-    {
-      municipio: "JERIQUARA",
-      abstencao_medio_completo: 17.2727,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_medio_completo: 17.8571,
-    },
-    {
-      municipio: "GASTÃO VIDIGAL",
-      abstencao_medio_completo: 18.1818,
-    },
-    {
-      municipio: "GUARANI D'OESTE",
-      abstencao_medio_completo: 18.5185,
-    },
-  ],
-  max_abstencoes_medio_completo: [
-    {
-      municipio: "LINS",
-      abstencao_medio_completo: 85.0237,
-    },
-    {
-      municipio: "JALES",
-      abstencao_medio_completo: 84.6617,
-    },
-    {
-      municipio: "PIEDADE",
-      abstencao_medio_completo: 82.8142,
-    },
-    {
-      municipio: "TUPÃ",
-      abstencao_medio_completo: 82.4598,
-    },
-    {
-      municipio: "PIRAPOZINHO",
-      abstencao_medio_completo: 81.0345,
-    },
-  ],
-  min_abstencoes_superior_completo: [
-    {
-      municipio: "ASPÁSIA",
-      abstencao_superior_completo: 17.1053,
-    },
-    {
-      municipio: "JERIQUARA",
-      abstencao_superior_completo: 17.2727,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_superior_completo: 17.8571,
-    },
-    {
-      municipio: "GASTÃO VIDIGAL",
-      abstencao_superior_completo: 18.1818,
-    },
-    {
-      municipio: "GUARANI D'OESTE",
-      abstencao_superior_completo: 18.5185,
-    },
-  ],
-  max_abstencoes_superior_completo: [
-    {
-      municipio: "LINS",
-      abstencao_superior_completo: 85.0237,
-    },
-    {
-      municipio: "JALES",
-      abstencao_superior_completo: 84.6617,
-    },
-    {
-      municipio: "PIEDADE",
-      abstencao_superior_completo: 82.8142,
-    },
-    {
-      municipio: "TUPÃ",
-      abstencao_superior_completo: 82.4598,
-    },
-    {
-      municipio: "PIRAPOZINHO",
-      abstencao_superior_completo: 81.0345,
-    },
-  ],
-  min_abstencoes_casados: [
-    {
-      municipio: "PIQUEROBI",
-      abstencao_casados: 7.2993,
-    },
-    {
-      municipio: "PEDRANÓPOLIS",
-      abstencao_casados: 7.933,
-    },
-    {
-      municipio: "NOVA CASTILHO",
-      abstencao_casados: 8.0292,
-    },
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_casados: 9.0791,
-    },
-    {
-      municipio: "JERIQUARA",
-      abstencao_casados: 9.6078,
-    },
-  ],
-  max_abstencoes_casados: [
-    {
-      municipio: "ILHA SOLTEIRA",
-      abstencao_casados: 36.4173,
-    },
-    {
-      municipio: "MOCOCA",
-      abstencao_casados: 36.1237,
-    },
-    {
-      municipio: "JUQUIÁ",
-      abstencao_casados: 36.0239,
-    },
-    {
-      municipio: "BARRETOS",
-      abstencao_casados: 35.8971,
-    },
-    {
-      municipio: "ROSANA",
-      abstencao_casados: 35.8257,
-    },
-  ],
-  min_abstencoes_solteiros: [
-    {
-      municipio: "SÃO JOÃO DE IRACEMA",
-      abstencao_solteiros: 6.1972,
-    },
-    {
-      municipio: "PIQUEROBI",
-      abstencao_solteiros: 6.7906,
-    },
-    {
-      municipio: "PRACINHA",
-      abstencao_solteiros: 6.8392,
-    },
-    {
-      municipio: "TACIBA",
-      abstencao_solteiros: 7.1493,
-    },
-    {
-      municipio: "PEDRANÓPOLIS",
-      abstencao_solteiros: 7.3128,
-    },
-  ],
-  max_abstencoes_solteiros: [
-    {
-      municipio: "SANTOS",
-      abstencao_solteiros: 35.2182,
-    },
-    {
-      municipio: "RIBEIRÃO PRETO",
-      abstencao_solteiros: 33.4838,
-    },
-    {
-      municipio: "PIRACICABA",
-      abstencao_solteiros: 32.5219,
-    },
-    {
-      municipio: "CAMPINAS",
-      abstencao_solteiros: 31.7989,
-    },
-    {
-      municipio: "ADAMANTINA",
-      abstencao_solteiros: 31.6388,
-    },
-  ],
-};
-module.exports = data;
