--- src/controllers/eleitorado_atualizado_json.js
@@ -1,432 +1,0 @@
-const data = {
-  eleitorado: [
-    {
-      municipio: "SÃO PAULO",
-      QT_ELEITORES_PERFIL: 8975761,
-      QT_ELEITORES_INC_NM_SOCIAL: 1195,
-      faixa_etaria: [
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "100 anos ou mais              ",
-          soma_eleitores_perfil: 8057,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "16 anos                       ",
-          soma_eleitores_perfil: 1891,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "17 anos                       ",
-          soma_eleitores_perfil: 16184,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "18 anos                       ",
-          soma_eleitores_perfil: 61493,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "19 anos                       ",
-          soma_eleitores_perfil: 112198,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "20 anos                       ",
-          soma_eleitores_perfil: 148419,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "21 a 24 anos                  ",
-          soma_eleitores_perfil: 683861,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "25 a 29 anos                  ",
-          soma_eleitores_perfil: 837902,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "30 a 34 anos                  ",
-          soma_eleitores_perfil: 895548,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "35 a 39 anos                  ",
-          soma_eleitores_perfil: 978109,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "40 a 44 anos                  ",
-          soma_eleitores_perfil: 928104,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "45 a 49 anos                  ",
-          soma_eleitores_perfil: 826670,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "50 a 54 anos                  ",
-          soma_eleitores_perfil: 754902,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "55 a 59 anos                  ",
-          soma_eleitores_perfil: 701904,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "60 a 64 anos                  ",
-          soma_eleitores_perfil: 600431,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "65 a 69 anos                  ",
-          soma_eleitores_perfil: 475047,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "70 a 74 anos                  ",
-          soma_eleitores_perfil: 340427,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "75 a 79 anos                  ",
-          soma_eleitores_perfil: 227820,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "80 a 84 anos                  ",
-          soma_eleitores_perfil: 157517,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "85 a 89 anos                  ",
-          soma_eleitores_perfil: 96153,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "90 a 94 anos                  ",
-          soma_eleitores_perfil: 71676,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "95 a 99 anos                  ",
-          soma_eleitores_perfil: 51350,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_faixa_etaria: "Inválido                      ",
-          soma_eleitores_perfil: 98,
-        },
-      ],
-      estado_civil: [
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "CASADO",
-          soma_eleitores_perfil: 2964504,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "DIVORCIADO",
-          soma_eleitores_perfil: 369899,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "NÃO INFORMADO",
-          soma_eleitores_perfil: 8596,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "SEPARADO JUDICIALMENTE",
-          soma_eleitores_perfil: 103420,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "SOLTEIRO",
-          soma_eleitores_perfil: 5305814,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_estado_civil: "VIÚVO",
-          soma_eleitores_perfil: 223528,
-        },
-      ],
-      grau_escolaridade: [
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ANALFABETO",
-          soma_eleitores_perfil: 174725,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL COMPLETO",
-          soma_eleitores_perfil: 662425,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL INCOMPLETO",
-          soma_eleitores_perfil: 1580046,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO MÉDIO COMPLETO",
-          soma_eleitores_perfil: 2512419,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "ENSINO MÉDIO INCOMPLETO",
-          soma_eleitores_perfil: 1617912,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "LÊ E ESCREVE",
-          soma_eleitores_perfil: 374111,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "SUPERIOR COMPLETO",
-          soma_eleitores_perfil: 1442597,
-        },
-        {
-          municipio: "SÃO PAULO",
-          desc_grau_escolaridade: "SUPERIOR INCOMPLETO",
-          soma_eleitores_perfil: 611526,
-        },
-      ],
-    },
-    {
-      municipio: "SÃO JOSÉ DOS CAMPOS",
-      QT_ELEITORES_PERFIL: 475288,
-      QT_ELEITORES_INC_NM_SOCIAL: 50,
-      faixa_etaria: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "16 anos                       ",
-          soma_eleitores_perfil: 80,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "17 anos                       ",
-          soma_eleitores_perfil: 1226,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "18 anos                       ",
-          soma_eleitores_perfil: 4536,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "19 anos                       ",
-          soma_eleitores_perfil: 7488,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "20 anos                       ",
-          soma_eleitores_perfil: 9403,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "21 a 24 anos                  ",
-          soma_eleitores_perfil: 39672,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "25 a 29 anos                  ",
-          soma_eleitores_perfil: 45339,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "30 a 34 anos                  ",
-          soma_eleitores_perfil: 49393,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "35 a 39 anos                  ",
-          soma_eleitores_perfil: 54230,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "40 a 44 anos                  ",
-          soma_eleitores_perfil: 51476,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "45 a 49 anos                  ",
-          soma_eleitores_perfil: 44899,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "50 a 54 anos                  ",
-          soma_eleitores_perfil: 41370,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "55 a 59 anos                  ",
-          soma_eleitores_perfil: 39339,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "60 a 64 anos                  ",
-          soma_eleitores_perfil: 34430,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "65 a 69 anos                  ",
-          soma_eleitores_perfil: 27209,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "70 a 74 anos                  ",
-          soma_eleitores_perfil: 15661,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "75 a 79 anos                  ",
-          soma_eleitores_perfil: 6261,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "80 a 84 anos                  ",
-          soma_eleitores_perfil: 2325,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "85 a 89 anos                  ",
-          soma_eleitores_perfil: 736,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "90 a 94 anos                  ",
-          soma_eleitores_perfil: 177,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "95 a 99 anos                  ",
-          soma_eleitores_perfil: 36,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_faixa_etaria: "Inválido                      ",
-          soma_eleitores_perfil: 2,
-        },
-      ],
-      estado_civil: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "CASADO",
-          soma_eleitores_perfil: 226716,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "DIVORCIADO",
-          soma_eleitores_perfil: 32435,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "SEPARADO JUDICIALMENTE",
-          soma_eleitores_perfil: 7835,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "SOLTEIRO",
-          soma_eleitores_perfil: 193245,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_estado_civil: "VIÚVO",
-          soma_eleitores_perfil: 15057,
-        },
-      ],
-      grau_escolaridade: [
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ANALFABETO",
-          soma_eleitores_perfil: 3727,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL COMPLETO",
-          soma_eleitores_perfil: 31599,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO FUNDAMENTAL INCOMPLETO",
-          soma_eleitores_perfil: 62961,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO MÉDIO COMPLETO",
-          soma_eleitores_perfil: 187746,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "ENSINO MÉDIO INCOMPLETO",
-          soma_eleitores_perfil: 40234,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "LÊ E ESCREVE",
-          soma_eleitores_perfil: 6451,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "SUPERIOR COMPLETO",
-          soma_eleitores_perfil: 110758,
-        },
-        {
-          municipio: "SÃO JOSÉ DOS CAMPOS",
-          desc_grau_escolaridade: "SUPERIOR INCOMPLETO",
-          soma_eleitores_perfil: 31812,
-        },
-      ],
-    },
-  ],
-  max_eleitorado_jovens: [
-    {
-      municipio: "ENGENHEIRO COELHO",
-      porcentagem_eleitorado_jovens: 33.1,
-    },
-    {
-      municipio: "IARAS",
-      porcentagem_eleitorado_jovens: 31.81,
-    },
-    {
-      municipio: "BOM SUCESSO DE ITARARÉ",
-      porcentagem_eleitorado_jovens: 31.38,
-    },
-    {
-      municipio: "NOVA CAMPINA",
-      porcentagem_eleitorado_jovens: 30.46,
-    },
-    {
-      municipio: "FRANCISCO MORATO",
-      porcentagem_eleitorado_jovens: 29.82,
-    },
-  ],
-  min_eleitorado_jovens: [
-    {
-      municipio: "SANTOS",
-      porcentagem_eleitorado_jovens: 16.88,
-    },
-    {
-      municipio: "SANTANA DA PONTE PENSA",
-      porcentagem_eleitorado_jovens: 16.91,
-    },
-    {
-      municipio: "SANTA RITA DO PASSA QUATRO",
-      porcentagem_eleitorado_jovens: 17.07,
-    },
-    {
-      municipio: "INDIAPORÃ",
-      porcentagem_eleitorado_jovens: 17.2,
-    },
-    {
-      municipio: "NEVES PAULISTA",
-      porcentagem_eleitorado_jovens: 17.21,
-    },
-  ],
-};
-
-module.exports = data;
