+++ src/controllers/eleitorado_json.js
@@ -387,352 +387,432 @@ const data = {
     {
       municipio: "ENGENHEIRO COELHO",
       porcentagem_eleitorado_jovens: 33.1,
+      valor_eleitorado_jovens: 4336,
     },
     {
       municipio: "IARAS",
       porcentagem_eleitorado_jovens: 31.81,
+      valor_eleitorado_jovens: 1274,
     },
     {
       municipio: "BOM SUCESSO DE ITARARÉ",
       porcentagem_eleitorado_jovens: 31.38,
+      valor_eleitorado_jovens: 909,
     },
     {
       municipio: "NOVA CAMPINA",
       porcentagem_eleitorado_jovens: 30.46,
+      valor_eleitorado_jovens: 1922,
     },
     {
       municipio: "FRANCISCO MORATO",
       porcentagem_eleitorado_jovens: 29.82,
+      valor_eleitorado_jovens: 30756,
     },
   ],
   min_eleitorado_jovens: [
     {
       municipio: "SANTOS",
       porcentagem_eleitorado_jovens: 16.88,
+      valor_eleitorado_jovens: 57529,
     },
     {
       municipio: "SANTANA DA PONTE PENSA",
       porcentagem_eleitorado_jovens: 16.91,
+      valor_eleitorado_jovens: 250,
     },
     {
       municipio: "SANTA RITA DO PASSA QUATRO",
       porcentagem_eleitorado_jovens: 17.07,
+      valor_eleitorado_jovens: 3336,
     },
     {
       municipio: "INDIAPORÃ",
       porcentagem_eleitorado_jovens: 17.2,
+      valor_eleitorado_jovens: 491,
     },
     {
       municipio: "NEVES PAULISTA",
       porcentagem_eleitorado_jovens: 17.21,
+      valor_eleitorado_jovens: 1079,
     },
   ],
   max_eleitorado_adultos: [
     {
       municipio: "HOLAMBRA",
       porcentagem_eleitorado_adultos: 61.43,
+      valor_eleitorado_adultos: 6403,
     },
     {
       municipio: "BARUERI",
       porcentagem_eleitorado_adultos: 61.22,
+      valor_eleitorado_adultos: 160266,
     },
     {
       municipio: "IGARAÇU DO TIETÊ",
       porcentagem_eleitorado_adultos: 60.93,
+      valor_eleitorado_adultos: 10107,
     },
     {
       municipio: "RAFARD",
       porcentagem_eleitorado_adultos: 60.91,
+      valor_eleitorado_adultos: 3862,
     },
     {
       municipio: "ILHABELA",
       porcentagem_eleitorado_adultos: 60.89,
+      valor_eleitorado_adultos: 14278,
     },
   ],
   min_eleitorado_adultos: [
     {
       municipio: "SANTA RITA D'OESTE",
       porcentagem_eleitorado_adultos: 50.4,
+      valor_eleitorado_adultos: 1133,
     },
     {
       municipio: "NOVA CANAÃ PAULISTA",
       porcentagem_eleitorado_adultos: 50.51,
+      valor_eleitorado_adultos: 946,
     },
     {
       municipio: "NOVA GUATAPORANGA",
       porcentagem_eleitorado_adultos: 50.79,
+      valor_eleitorado_adultos: 930,
     },
     {
       municipio: "TURMALINA",
       porcentagem_eleitorado_adultos: 50.81,
+      valor_eleitorado_adultos: 848,
     },
     {
       municipio: "SANTANA DA PONTE PENSA",
       porcentagem_eleitorado_adultos: 50.81,
+      valor_eleitorado_adultos: 751,
     },
   ],
   max_eleitorado_idosos: [
     {
       municipio: "SANTANA DA PONTE PENSA",
       porcentagem_eleitorado_idosos: 32.27,
+      valor_eleitorado_idosos: 477,
     },
     {
       municipio: "NOVA CANAÃ PAULISTA",
       porcentagem_eleitorado_idosos: 31.82,
+      valor_eleitorado_idosos: 596,
     },
     {
       municipio: "SANTA RITA D'OESTE",
       porcentagem_eleitorado_idosos: 31.45,
+      valor_eleitorado_idosos: 707,
     },
     {
       municipio: "ÁGUAS DE SÃO PEDRO",
       porcentagem_eleitorado_idosos: 31.33,
+      valor_eleitorado_idosos: 1193,
     },
     {
       municipio: "TURMALINA",
       porcentagem_eleitorado_idosos: 31.16,
+      valor_eleitorado_idosos: 520,
     },
   ],
   min_eleitorado_idosos: [
     {
       municipio: "ENGENHEIRO COELHO",
       porcentagem_eleitorado_idosos: 12.28,
+      valor_eleitorado_idosos: 1609,
     },
     {
       municipio: "CAJAMAR",
       porcentagem_eleitorado_idosos: 12.48,
+      valor_eleitorado_idosos: 6637,
     },
     {
       municipio: "BARRINHA",
       porcentagem_eleitorado_idosos: 12.99,
+      valor_eleitorado_idosos: 2667,
     },
     {
       municipio: "ARAÇARIGUAMA",
       porcentagem_eleitorado_idosos: 13.05,
+      valor_eleitorado_idosos: 1897,
     },
     {
       municipio: "PONTAL",
       porcentagem_eleitorado_idosos: 13.34,
+      valor_eleitorado_idosos: 3350,
     },
   ],
   max_eleitorado_superior_completo: [
     {
       municipio: "SÃO CAETANO DO SUL",
       porcentagem_eleitorado_superior_completo: 41.16,
+      valor_eleitorado_superior_completo: 48300,
     },
     {
       municipio: "VALINHOS",
       porcentagem_eleitorado_superior_completo: 28.57,
+      valor_eleitorado_superior_completo: 21574,
     },
     {
       municipio: "ÁGUAS DE SÃO PEDRO",
       porcentagem_eleitorado_superior_completo: 28.05,
+      valor_eleitorado_superior_completo: 1068,
     },
     {
       municipio: "ASSIS",
       porcentagem_eleitorado_superior_completo: 24.56,
+      valor_eleitorado_superior_completo: 15474,
     },
     {
       municipio: "PRESIDENTE PRUDENTE",
       porcentagem_eleitorado_superior_completo: 24.38,
+      valor_eleitorado_superior_completo: 36699,
     },
   ],
   min_eleitorado_superior_completo: [
     {
       municipio: "CABREÚVA",
       porcentagem_eleitorado_superior_completo: 3.14,
+      valor_eleitorado_superior_completo: 1053,
     },
     {
       municipio: "ITAPEVI",
       porcentagem_eleitorado_superior_completo: 4.18,
+      valor_eleitorado_superior_completo: 6420,
     },
     {
       municipio: "RIBEIRÃO GRANDE",
       porcentagem_eleitorado_superior_completo: 4.34,
+      valor_eleitorado_superior_completo: 250,
     },
     {
       municipio: "RIBEIRÃO CORRENTE",
       porcentagem_eleitorado_superior_completo: 4.43,
+      valor_eleitorado_superior_completo: 183,
     },
     {
       municipio: "AVAÍ",
       porcentagem_eleitorado_superior_completo: 4.44,
+      valor_eleitorado_superior_completo: 178,
     },
   ],
   max_eleitorado_medio_completo: [
     {
       municipio: "RIO GRANDE DA SERRA",
       porcentagem_eleitorado_medio_completo: 45.52,
+      valor_eleitorado_medio_completo: 12982,
     },
     {
       municipio: "ALUMÍNIO",
       porcentagem_eleitorado_medio_completo: 44.53,
+      valor_eleitorado_medio_completo: 5405,
     },
     {
       municipio: "POÁ",
       porcentagem_eleitorado_medio_completo: 44.4,
+      valor_eleitorado_medio_completo: 31728,
     },
     {
       municipio: "PIQUETE",
       porcentagem_eleitorado_medio_completo: 43.52,
+      valor_eleitorado_medio_completo: 4238,
     },
     {
       municipio: "HORTOLÂNDIA",
       porcentagem_eleitorado_medio_completo: 43.39,
+      valor_eleitorado_medio_completo: 54100,
     },
   ],
   min_eleitorado_medio_completo: [
     {
       municipio: "RIBEIRÃO CORRENTE",
       porcentagem_eleitorado_medio_completo: 19.14,
+      valor_eleitorado_medio_completo: 790,
     },
     {
       municipio: "UCHOA",
       porcentagem_eleitorado_medio_completo: 19.26,
+      valor_eleitorado_medio_completo: 1537,
     },
     {
       municipio: "GUAPIAÇU",
       porcentagem_eleitorado_medio_completo: 19.84,
+      valor_eleitorado_medio_completo: 2941,
     },
     {
       municipio: "NOVA ALIANÇA",
       porcentagem_eleitorado_medio_completo: 20.07,
+      valor_eleitorado_medio_completo: 1013,
     },
     {
       municipio: "SÃO JOSÉ DA BELA VISTA",
       porcentagem_eleitorado_medio_completo: 20.16,
+      valor_eleitorado_medio_completo: 1381,
     },
   ],
   max_eleitorado_analfabeto: [
     {
       municipio: "LAGOINHA",
       porcentagem_eleitorado_analfabeto: 9.87,
+      valor_eleitorado_analfabeto: 508,
     },
     {
       municipio: "SÃO LUÍS DO PARAITINGA",
       porcentagem_eleitorado_analfabeto: 9.81,
+      valor_eleitorado_analfabeto: 954,
     },
     {
       municipio: "BARRA DO TURVO",
       porcentagem_eleitorado_analfabeto: 9.74,
+      valor_eleitorado_analfabeto: 555,
     },
     {
       municipio: "RIBEIRÃO BRANCO",
       porcentagem_eleitorado_analfabeto: 9.48,
+      valor_eleitorado_analfabeto: 1268,
     },
     {
       municipio: "PEDRA BELA",
       porcentagem_eleitorado_analfabeto: 9.21,
+      valor_eleitorado_analfabeto: 397,
     },
   ],
   min_eleitorado_analfabeto: [
     {
       municipio: "SÃO CAETANO DO SUL",
       porcentagem_eleitorado_analfabeto: 0.12,
+      valor_eleitorado_analfabeto: 141,
     },
     {
       municipio: "IRACEMÁPOLIS",
       porcentagem_eleitorado_analfabeto: 0.42,
+      valor_eleitorado_analfabeto: 65,
     },
     {
       municipio: "LINS",
       porcentagem_eleitorado_analfabeto: 0.55,
+      valor_eleitorado_analfabeto: 267,
     },
     {
       municipio: "VALINHOS",
       porcentagem_eleitorado_analfabeto: 0.59,
+      valor_eleitorado_analfabeto: 443,
     },
     {
       municipio: "LORENA",
       porcentagem_eleitorado_analfabeto: 0.64,
+      valor_eleitorado_analfabeto: 337,
     },
   ],
   max_eleitorado_casados: [
     {
       municipio: "MARAPOAMA",
       porcentagem_eleitorado_casados: 55.77,
+      valor_eleitorado_casados: 1420,
     },
     {
       municipio: "CATIGUÁ",
       porcentagem_eleitorado_casados: 55.17,
+      valor_eleitorado_casados: 2872,
     },
     {
       municipio: "SANTA SALETE",
       porcentagem_eleitorado_casados: 54.76,
+      valor_eleitorado_casados: 885,
     },
     {
       municipio: "BORÁ",
       porcentagem_eleitorado_casados: 54.27,
+      valor_eleitorado_casados: 528,
     },
     {
       municipio: "ALFREDO MARCONDES",
       porcentagem_eleitorado_casados: 54.06,
+      valor_eleitorado_casados: 1885,
     },
   ],
   min_eleitorado_casados: [
     {
       municipio: "SÃO VICENTE",
       porcentagem_eleitorado_casados: 28.34,
+      valor_eleitorado_casados: 71229,
     },
     {
       municipio: "GUARUJÁ",
       porcentagem_eleitorado_casados: 28.75,
+      valor_eleitorado_casados: 64672,
     },
     {
       municipio: "BERTIOGA",
       porcentagem_eleitorado_casados: 30.38,
+      valor_eleitorado_casados: 13592,
     },
     {
       municipio: "ITAPEVI",
       porcentagem_eleitorado_casados: 30.67,
+      valor_eleitorado_casados: 47151,
     },
     {
       municipio: "FRANCO DA ROCHA",
       porcentagem_eleitorado_casados: 31.46,
+      valor_eleitorado_casados: 31506,
     },
   ],
   max_eleitorado_solteiros: [
     {
       municipio: "GUARUJÁ",
       porcentagem_eleitorado_solteiros: 64.62,
+      valor_eleitorado_solteiros: 145342,
     },
     {
       municipio: "SÃO VICENTE",
       porcentagem_eleitorado_solteiros: 63.74,
+      valor_eleitorado_solteiros: 160224,
     },
     {
       municipio: "ITAPEVI",
       porcentagem_eleitorado_solteiros: 63.46,
+      valor_eleitorado_solteiros: 97573,
     },
     {
       municipio: "CABREÚVA",
       porcentagem_eleitorado_solteiros: 62.82,
+      valor_eleitorado_solteiros: 21094,
     },
     {
       municipio: "BERTIOGA",
       porcentagem_eleitorado_solteiros: 62.19,
+      valor_eleitorado_solteiros: 27823,
     },
   ],
   min_eleitorado_solteiros: [
     {
       municipio: "MARAPOAMA",
       porcentagem_eleitorado_solteiros: 33.27,
+      valor_eleitorado_solteiros: 847,
     },
     {
       municipio: "SANTA SALETE",
       porcentagem_eleitorado_solteiros: 33.29,
+      valor_eleitorado_solteiros: 538,
     },
     {
       municipio: "SANTANA DA PONTE PENSA",
       porcentagem_eleitorado_solteiros: 33.42,
+      valor_eleitorado_solteiros: 494,
     },
     {
       municipio: "RUBINÉIA",
       porcentagem_eleitorado_solteiros: 34.09,
+      valor_eleitorado_solteiros: 956,
     },
     {
       municipio: "ELISIÁRIO",
       porcentagem_eleitorado_solteiros: 34.12,
+      valor_eleitorado_solteiros: 914,
     },
   ],
 };
